sourced out pid generation, as it allows for more structure now.

+++ src/corrections/higherorder.cpp
@@ -8,7 +8,7 @@
 using namespace std;
 
 void higherorder::innerloop(treereader *tree, std::vector<std::atomic<bool>>
-        &goodevents, std::vector<int> range) {
+        &goodevents, std::vector<uint> range) {
     // Step 1: Cloning histograms
     vector<vector<vector<TH2D>>> _culpritdiag;
     for(auto &i: culpritdiag){
@@ -154,8 +154,8 @@ void higherorder::analyse(const std::vector<std::string> input, TFile *output) {
 
     vector<thread> th;
     for(uint i=0; i<threads; i++){
-        vector<int> ranges = {i*goodevents.size()/threads,
-                              (i+1)*goodevents.size()/threads-1};
+        vector<uint> ranges = {(uint)(i*goodevents.size()/threads),
+                              (uint)((i+1)*goodevents.size()/threads-1)};
         th.emplace_back(thread(&higherorder::innerloop, this, tree.at(i),
                                ref(goodevents),ranges));
     }
