+++ src/cutclasses/ICcut.cpp
@@ -7,7 +7,7 @@
 using namespace std;
 
 void iccut::innerloop(treereader *tree, std::vector<std::atomic<bool>>
-                        &goodevents, std::vector<int> range) {
+                        &goodevents, std::vector<uint> range) {
     //Step 1: Cloning histograms
     vector<vector<TH2D>> _comparediag;
     for(auto &i: comparediag){
@@ -99,8 +99,8 @@ void iccut::analyse(const std::vector<std::string> input, TFile *output) {
 
     vector<thread> th;
     for(uint i=0; i<threads; i++){
-        vector<int> ranges = {i*goodevents.size()/threads,
-                              (i+1)*goodevents.size()/threads-1};
+        vector<uint> ranges = {(uint)(i*goodevents.size()/threads),
+                               (uint)((i+1)*goodevents.size()/threads-1)};
         th.emplace_back(thread(&iccut::innerloop, this, tree.at(i),
                                ref(goodevents),ranges));
     }
